requirements.txt
pandas
matplotlib
scikit-learn
python-telegram-bot==20.3
plotly